+++ src/main/java/net/wizards/WizardsMod.java
@@ -20,7 +20,7 @@ public class WizardsMod {
     public static final String ID = "wizards";
 
     public static ConfigManager<ItemConfig> itemConfig = new ConfigManager<ItemConfig>
-            ("items_v2", Default.itemConfig)
+            ("items_v3", Default.itemConfig)
             .builder()
             .setDirectory(ID)
             .sanitize(true)